faktur-service/requirements.txt
# PRODUCTION RAILWAY DEPLOYMENT
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
